[APPLE-2367] Address warnings

+++ APCAppCore/APCAppCore/Library/Objects/APCSegmentedButton.m
@@ -63,7 +63,7 @@
 
     for (NSUInteger i=0; i<self.buttons.count; i++) {
         UIButton *button = self.buttons[i];
-        if (i == selectedIndex) {
+        if (i == (NSUInteger)selectedIndex) {
             button.tintColor = self.highlightColor;
         } else {
             button.tintColor = self.normalColor;
